third_party/blink/web_tests/editing/shadow/shadow-selection-not-exported-expected.txt

PASS assertNotInShadow(selection.anchorNode) is true
PASS assertNotInShadow(selection.focusNode) is true
PASS assertNotInShadow(selection.anchorNode) is true
PASS assertNotInShadow(selection.focusNode) is true
PASS assertNotInShadow(selection.anchorNode) is true
PASS assertNotInShadow(selection.focusNode) is true
PASS successfullyParsed is true

TEST COMPLETE


This test ensures shadow root won't expose to JS layer.

BEFORE
AFTER
